Made email sending synchronous
